WIP: restructured code; put everything in the .api namespace; updated tests code
